refactor(extractnoabstractiondirs): extractNoAbstractionDirs を別ファイルに切り出し
